+++ components/AppFooter.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts"></script>
 
 <template>
-  <footer class="font-inter bg-[#101113] py-4 text-center text-sm border-t border-gray-700 text-white">
+  <footer class="font-inter bg-black py-4 text-center text-sm border-t border-gray-700 text-white">
     <div class="flex flex-col gap-3 text-gray-400">
       <p class="text-lg font-bold">RIZQYA <br> ADZRA.</p>
       <p>{{ $t('footer_1') }}</p>