+++ components/LanguageSwitcher.vue
@@ -15,7 +15,7 @@ const changeLanguage = (code) => {
   <div class="relative inline-block text-left">
     <button
       @click="isOpen = !isOpen"
-      class="flex items-center gap-2 px-3 py-2 bg-[#2b2b2b] text-[#f5f0e6] rounded-full hover:opacity-90 focus:outline-none transition"
+      class="flex items-center gap-2 px-3 py-2 bg-slate-200 dark:bg-[#2b2b2b] text-black dark:text-white rounded-full hover:opacity-90 focus:outline-none transition duration-500"
     >
       <Icon name="ph:globe-bold" class="text-3xl" />
       <Icon name="ph:caret-down-bold" class="text-xs" />
@@ -23,19 +23,20 @@ const changeLanguage = (code) => {
 
     <div
       v-if="isOpen"
-      class="absolute right-0 mt-2 w-32 bg-[#2b2b2b] text-white rounded-md shadow-lg z-50"
+      class="absolute right-0 mt-2 w-32 bg-slate-200 text-black dark:bg-[#2b2b2b] dark:text-white rounded-md shadow-lg z-50"
     >
       <ul class="py-1 text-sm">
         <li
           v-for="l in locales"
           :key="l.code"
           @click="changeLanguage(l.code)"
-          class="px-2 py-2 mx-3 my-1 rounded hover:bg-[#444] cursor-pointer font-bold"
-          :class="locale === l.code ? 'bg-[#444]' : ''"
+          class="px-2 py-2 mx-3 my-1 rounded cursor-pointer font-bold hover:bg-slate-500 hover:text-white dark:hover:bg-[#444]"
+          :class="locale === l.code ? 'bg-slate-500 dark:bg-[#444] text-white' : ''"
         >
           {{ l.name }}
         </li>
       </ul>
+
     </div>
   </div>
 </template>