+++ components/VideoPlay.vue
@@ -1,6 +1,5 @@
 <template>
   <div class="relative w-full overflow-hidden bg-black">
-    <!-- Wrapper untuk ratio 21:9 -->
     <div class="relative w-full" style="aspect-ratio: 23 / 9;">
       <video
         autoplay
@@ -14,10 +13,8 @@
         Your browser does not support the video tag.
       </video>
 
-      <!-- Overlay -->
-      <div class="absolute inset-0 bg-black/60 z-10"></div>
+      <div class="absolute inset-0 bg-black/50 z-10"></div>
 
-      <!-- Text -->
       <div class="absolute inset-0 flex flex-col gap-3 items-center justify-center z-20 text-white -translate-y-11">
         <p class="font-jaro text-8xl font-bold">{{ $t('welcome') }}</p>
         <p class="font-inter font-bold text-xs">{{ $t('introduction') }}</p>