+++ components/base/NavigationButton.vue
@@ -1,49 +1,30 @@
 <template>
-  <button 
-    class="font-inter text-white border-b-[6px] p-2 transition-all border-transparent"
+  <button
+    @click="navigateTo(link)"
     :class="[
-      'hover:border-[#714AFF]', 
-      { 
-        'border-[#714AFF]': isActive, 
-      }
+      'relative p-4 transition-all rounded after:absolute after:left-0 after:bottom-0 after:h-[6px] after:rounded-lg after:transition-all duration-500',
+      route.path === link
+        ? 'after:w-full after:bg-[#714AFF]'
+        : 'after:w-0 after:bg-[#714AFF] hover:after:w-full'
     ]"
-    @click="handleClick">
+  >
     {{ buttonName }}
   </button>
 </template>
 
-<script>
-import { computed } from "vue";
-import { useRoute } from "vue-router";
+<script setup>
+import { useRoute, navigateTo } from '#imports'
 
-export default {
-  props: {
-    buttonName: {
-      type: String,
-      default: "NAVIGATION BUTTON"
-    },
-    onClick: {
-      type: Function,
-      required: true,
-    },
-    path: {
-      type: String, 
-      required: true,
-    }
+const props = defineProps({
+  buttonName: {
+    type: String,
+    required: true
   },
-  setup(props) {
-    const route = useRoute();
-
-    const isActive = computed(() => route.path === props.path);
-
-    return { isActive };
-  },
-  methods: {
-    handleClick() {
-      if (this.onClick) {
-        this.onClick()
-      }
-    }
+  link: {
+    type: String,
+    required: true
   }
-}
+})
+
+const route = useRoute()
 </script>